BEST threasholds: 02 - 0.66 05 - 0.61 001 - 0.75 0005 - 0.76 0001 - 0.67

+++ utils/remove_hair.py
@@ -1,9 +1,12 @@
 import os
 import numpy as np
 import cv2
+import multiprocessing as mp
+from pathlib import Path
 
 IN_DIR = '/media/3tstor/ml/IdeaProjects/ml_kaggle_competition1/input/jpeg/train/'
-OUT_DIR = '/media/3tstor/ml/IdeaProjects/ml_kaggle_competition1/input/jpeg/train_nohair/'
+OUT_DIR = '/home/werdn/input/jpeg/train_nohair/'
+
 
 def hair_remove(image):
     # convert image to grayScale
@@ -24,15 +27,28 @@ def hair_remove(image):
     return final_image
 
 
-def main():
-    train_images_512 = [(IN_DIR + i, i) for i in os.listdir(IN_DIR)]
-    for i, image_file in enumerate(train_images_512):
+def process_one_img(image_file):
+    my_file = Path(OUT_DIR + image_file[1])
+    if my_file.is_file():
+        # file exists
+        print('file exists ' + image_file[1])
+        return
+    else:
         print('start process file ' + image_file[1])
         img = cv2.imread(image_file[0], cv2.IMREAD_COLOR)
         no_hair_img = hair_remove(img)
         cv2.imwrite(OUT_DIR + image_file[1], no_hair_img)
-        print('process file ' + image_file[1])
+        print('processed file ' + image_file[1])
 
+
+def main():
+    train_images_512 = [(IN_DIR + i, i) for i in os.listdir(IN_DIR)]
+    for i, image_file in enumerate(train_images_512):
+        process_one_img(image_file)
+    # imgs = [(IN_DIR + i, i) for i in os.listdir(IN_DIR)]
+    # pool = mp.Pool(processes=4)
+    # outputs = pool.map(process_one_img, imgs)
+    # pool.close()
 
 
 if __name__ == "__main__":
